Create middleware to fix Azure problem
Django-auditlog was causing issues with Azure environment. Problem was
that Azure reverse proxy forwards the end-users ip to HTTP_X_CLIENT_IP
request header, but Django-auditlog is expecting it to be in
HTTP_X_FORWARDED_FOR. We decided to create middleware to solve this
problem.

Simply put create middleware that inserts data from HTTP_X_CLIENT_IP to
HTTP_X_FORWARDED_FOR. We only want to use this middleware in Azure
environment, so initializing will raise MiddlewareNotUsed if
AZURE_DEPLOYMENT setting if False.

Refs: LIIK-62