Add function to link additional traffic signs to main traffic signs
There is no clear specification on how the additional traffic signs can be linked
to main traffic signs. So we have to come up with a solution with our own observation
with the data. This commit implement a solution can be described as follows:

1. Find all traffic signs with the type that starts with number 8. It seems Finnish
Transport Infrastructure Agency follows certain patterns to codify the traffic signs.
And additional traffic signs are all starts with 8, see details:
https://vayla.fi/web/en/road-network/traffic-signs/additional-panels

2. Find other traffic signs that are close the additional traffic signs (less
than 1m or maybe even less) on the plane (only check the x,y coordinates)

3. There might be multiple other traffic signs close by, if that's the case
we select the one that has bigger z-coordinate but closest to the additional
panel (meaning the traffic sign that appears immediately above the target
additional panel)

Refs: LIIK-70