Plan Admin UI
Implement basic Django admin interface for the new Plan model.

Add a custom view for managing the relations to Plan model as a sub-view
under the admin Plan change view. This is a view only for setting
the foreign key relations from one place. Do not include instances
that are related to some other plan in choices to avoid accidents.

Refs LIIK-17