Update fk on_delete rules to PROTECT where needed
Most foreign key relations had CASCADE on_delete rule. In the scope of
this system it is unwanted as it makes the platform prone to unwanted
data loss. Switch the on_delete rule to PROTECT so the deletion must
be done explicitly for all related models.

Note that the file model foreign key still have CASCADE as the on_delete
rule since deleting the individual file models separately doesn't really
make that much sense as the files are more or less equivalent to any
other regular attribute on the model.

Refs LIIK-136