Rename TrafficSignCode to TrafficControlDeviceType
TrafficSignCode as a model name was misleading since it was a lot more
than what the name suggested. Rename it to TrafficControlDeviceType
that represents the purpose of the model more accurately.

Rename all the related fields from "code" and "type" to "device_type"
for consitency with the updated name. Update admin, filter, serializer,
and viewset classes and url patterns to match the new model name.

Refs LIIK-136