Convert katajanokka importer to management command
Convert old katajanokka traffic sign importer script to a Django
management command.

Main improvements in using management command over the custom script is
GeoDjango utilities can be used to read the SHP file instead of relying
on osgeo. Also it's possible to write the traffic signs directly to the
database instead of doing POST requests to the API.

This in somewhat of a hurry, so there's no unit tests at this point.
Writing tests for this is on the backlog but at his point testing with
the real import data must suffice.

Refs LIIK-143