Show Enum logical value in admin select inputs
Show Enum's logical value after the label in in Django admin.

To achieve this implement a ModelAdmin mixin class that sets the form
field for every EnumField and EnumIntegerField to a custom field that
uses a widget which sets each option label to contain the logical value
in parenthesis after the label.

Refs LIIK-164