Enable hard delete in django admin
Previously Django admin was set to soft delete model instances if the
model was a soft deletable model. This was to be reverted because it
was thought that model instances should be deletable for real in django
admin. The API is still only able to soft delete.

However it should still be possible to do soft delete via Django admin.
Implement changelist actions to retain that feature with minimal effort.

Also implement a soft delete list filter that allows to list all soft
deleted instances and possibly recover them if necessary. Soft deleted
instances are hidden by default.

Refs LIIK-147