Add a map view to visualize traffic signs on the map
The map view can only be accessed by staff users, and the layers
available in map view can be defined in the admin page. The layer
identifier must match the layer name defined in GeoServer. Currently
we do not have all the layers created in GeoServer, so we introduce
a Layer model to allow admin adding layers to map view at a later
time without chaning the code.

Refs: LIIK-78