Owner model data migration
Implement data migration that intializes the database with following
predefined choices:

- City of Helsinki
- State
- Private
- Unknown

Assign Owner for all models based on the old owner text field that have
relation ship to it. If the owner text isn't close enough to any of the
existing choices a new Owner instance is created.

Add backwards migration that populates the owner char field with the
finnish name of the related Owner instance.

Refs LIIK-173