Add owner admin list filter for related models
Add owner to admin list filter for all models that have a relation to
owner model.

This done is in case there is some unwanted owner instances that
can not be deleted because some traffic control device has a relation
to the unwanted instance. Filtering list views by the unwanted owner
allows a lot less painful way to go through all related models so that
the relation to the unwanted instance can be replaced with correct
instance.

Refs LIIK-173