Provide multiple icons (SVG and PNG) for traffic control device type
Added script `svg_icons_to_png.py` to convert SVG icons to PNG icons in
traffic_control static assets. This probably was not the best way to
generate PNG icons build time, but it works now. The script utilizes
`cairosvg` library to conversion.

Traffic control device icons (SVG and PNGs) are now offered in `icons`
field with full URLs to the icons. The old `icon` field is still there
for backward compatibility. Also it's value is used in generating the
icon URLs.

Refs LIIK-546