WIP Fix the dropdown selector for the icon file
We can basically use most of what was available for
TrafficSign{Plan,Real}, except we're dealing with a different field name
(and slightly different structure), but template, javascript and CSS can
likely be reused.

There is a crash, both on the new thing and the old thing where if you
use the modal (from the + sign) in the field editor to create a new icon
to use (or traffic sign device type) the widget is unable to render the
new icon.

If nothing else, it would be good at least to handle the failure to find
the icon gracefully and tell the user "save and continue editing" to see
the updated icon.