feat: Separate icon storage, enable overwrite, enforce unique filename
We don't know yet how allowing overwrite will affect other workflows
yet, so we separate icon storage settings into from the default storage
settings for now.

Other changes on the related post_save signal handlers:

* Make use of the file's storage info instead of default_storage
* Don't make the PNG generation step exclusive to new files, or new PNG
  files won't be generated in case of SVG overwrite.

Other notes:

We have to handle in some manner the scenario where FooIcon gets created
with with "foo.svg" and then later a user attempts to overwrite it with
"bar.svg"

The issue with such scenario is that with the current code it would
leave "foo.svg" and its related "foo.png" files dangling in the storage,
so one approach to fixing or preventing that needs to be done.

* Prevent uploads with different name for an existing row, or
* Forcibly rename uploads with different name for an existing row, or
* Wipe old files as we're writing the new ones

Refs: LIIK-806